test: netaddress add more tests.
