+++ package.json
@@ -28,7 +28,7 @@
     "bfile": "~0.2.2",
     "bfilter": "~1.0.5",
     "bheep": "~0.1.5",
-    "binet": "~0.3.6",
+    "binet": "~0.3.7",
     "blgr": "~0.2.0",
     "blru": "~0.1.6",
     "blst": "~0.1.5",
